refactor(ui): split PlayerAvatar into modules + i18n the skin chip (ARC-755)
Break the 488-line PlayerAvatar into focused modules (constants, colors,
RaysHalo, AvatarDisc, CardChrome) under the 500-line cap; behavior and all
testids preserved. Lift the hardcoded "SKIN" chip label into i18n via a
consumer-supplied `prefix` (added to mannequin.stage in all five locales),
so packages/ui carries no user-facing literal.
