+++ apps/web/src/shared/i18n/messages/pages/shop/ru.ts
@@ -26,7 +26,7 @@ export const shopRu = {
   },
   mannequin: {
     tryOn: 'Примерка',
-    stage: { level: 'УР {level} · Онлайн', online: 'Онлайн' },
+    stage: { level: 'УР {level} · Онлайн', online: 'Онлайн', skin: 'Скин' },
     slots: {
       avatar: {
         label: 'Аватар',
